requirements.txt
python-binance
websockets
asyncio
redis
supabase
numpy
pandas
python-dotenv
